Refactoring and fixing Save team objects to use a mapper and fixing issues

+++ database/migrations/2025_02_14_205440_add_team_to_task_workflows.php
@@ -14,7 +14,9 @@ return new class extends Migration {
             if (!Schema::hasColumn('task_workflows', 'name')) {
                 $table->string('name')->default('')->after('id');
             }
-            $table->foreignId('team_id')->after('id')->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('task_workflows', 'team_id')) {
+                $table->foreignId('team_id')->after('id')->constrained()->onDelete('cascade');
+            }
             $table->string('description')->default('')->after('name');
         });
     }
